+++ mbf_abstract_nav/include/mbf_abstract_nav/abstract_action_base.hpp
@@ -36,34 +36,49 @@
  *
  */
 
-#ifndef MBF_ABSTRACT_NAV__ABSTRACT_ACTION_H_
-#define MBF_ABSTRACT_NAV__ABSTRACT_ACTION_H_
+#ifndef MBF_ABSTRACT_NAV__ABSTRACT_ACTION_BASE_H_
+#define MBF_ABSTRACT_NAV__ABSTRACT_ACTION_BASE_H_
 
 #include <actionlib/server/action_server.h>
-#include <mbf_abstract_nav/MoveBaseFlexConfig.h>
-#include "mbf_abstract_nav/robot_information.h"
+#include <mbf_utility/robot_information.h>
 
-namespace mbf_abstract_nav{
+#include "mbf_abstract_nav/MoveBaseFlexConfig.h"
+
+namespace mbf_abstract_nav
+{
 
 template <typename Action, typename Execution>
-class AbstractAction
+class AbstractActionBase
 {
  public:
-  typedef boost::shared_ptr<AbstractAction> Ptr;
+  typedef boost::shared_ptr<AbstractActionBase> Ptr;
   typedef typename actionlib::ActionServer<Action>::GoalHandle GoalHandle;
   typedef boost::function<void (GoalHandle &goal_handle, Execution &execution)> RunMethod;
   typedef struct{
     typename Execution::Ptr execution;
-    boost::thread* thread_ptr;
+    boost::thread* thread_ptr = NULL;
     GoalHandle goal_handle;
+    bool in_use = false;
   } ConcurrencySlot;
 
 
-  AbstractAction(
-      const std::string& name,
-      const RobotInformation &robot_info,
+  AbstractActionBase(
+      const std::string &name,
+      const mbf_utility::RobotInformation &robot_info,
       const RunMethod run_method
   ) : name_(name), robot_info_(robot_info), run_(run_method){}
+
+  virtual ~AbstractActionBase()
+  {
+    // cleanup threads used on executions
+    boost::lock_guard<boost::mutex> guard(slot_map_mtx_);
+    typename std::map<uint8_t, ConcurrencySlot>::iterator slot_it = concurrency_slots_.begin();
+    for (; slot_it != concurrency_slots_.end(); ++slot_it)
+    {
+      threads_.remove_thread(slot_it->second.thread_ptr);
+      delete slot_it->second.thread_ptr;
+    }
+  }
 
   virtual void start(
       GoalHandle &goal_handle,
@@ -76,56 +91,57 @@ class AbstractAction
     {
       goal_handle.setCanceled();
     }
-    else {
-      slot_map_mtx_.lock();
-      typename std::map<uint8_t, ConcurrencySlot>::iterator slot_it =
-          concurrency_slots_.find(slot);
-      slot_map_mtx_.unlock();
-      if (slot_it != concurrency_slots_.end()) {
-        // if there is a plugin running on the same slot, cancel it
+    else
+    {
+      boost::lock_guard<boost::mutex> guard(slot_map_mtx_);
+      typename std::map<uint8_t, ConcurrencySlot>::iterator slot_it = concurrency_slots_.find(slot);
+      if (slot_it != concurrency_slots_.end() && slot_it->second.in_use) {
+        // if there is already a plugin running on the same slot, cancel it
         slot_it->second.execution->cancel();
+
         if (slot_it->second.thread_ptr->joinable()) {
           slot_it->second.thread_ptr->join();
         }
       }
-      boost::lock_guard<boost::mutex> guard(slot_map_mtx_);
+
       // fill concurrency slot with the new goal handle, execution, and working thread
+      concurrency_slots_[slot].in_use = true;
       concurrency_slots_[slot].goal_handle = goal_handle;
       concurrency_slots_[slot].goal_handle.setAccepted();
       concurrency_slots_[slot].execution = execution_ptr;
-      concurrency_slots_[slot].thread_ptr = threads_.create_thread(boost::bind(
-          &AbstractAction::runAndCleanUp, this,
-          boost::ref(concurrency_slots_[slot].goal_handle), execution_ptr));
+      if (concurrency_slots_[slot].thread_ptr)
+      {
+        // cleanup previous execution; otherwise we will leak threads
+        threads_.remove_thread(concurrency_slots_[slot].thread_ptr);
+        delete concurrency_slots_[slot].thread_ptr;
+      }
+      concurrency_slots_[slot].thread_ptr =
+        threads_.create_thread(boost::bind(&AbstractActionBase::run, this, boost::ref(concurrency_slots_[slot])));
     }
   }
 
-  virtual void cancel(GoalHandle &goal_handle){
+  virtual void cancel(GoalHandle &goal_handle)
+  {
     uint8_t slot = goal_handle.getGoal()->concurrency_slot;
 
     boost::lock_guard<boost::mutex> guard(slot_map_mtx_);
     typename std::map<uint8_t, ConcurrencySlot>::iterator slot_it = concurrency_slots_.find(slot);
-    if(slot_it != concurrency_slots_.end())
+    if (slot_it != concurrency_slots_.end())
     {
       concurrency_slots_[slot].execution->cancel();
     }
   }
 
-  virtual void runAndCleanUp(GoalHandle &goal_handle, typename Execution::Ptr execution_ptr){
-    uint8_t slot = goal_handle.getGoal()->concurrency_slot;
-
-    execution_ptr->preRun();
-    run_(goal_handle, *execution_ptr);
+  virtual void run(ConcurrencySlot &slot)
+  {
+    slot.execution->preRun();
+    run_(slot.goal_handle, *slot.execution);
     ROS_DEBUG_STREAM_NAMED(name_, "Finished action \"" << name_ << "\" run method, waiting for execution thread to finish.");
-    execution_ptr->join();
-    ROS_DEBUG_STREAM_NAMED(name_, "Execution thread for action \"" << name_ << "\" stopped, cleaning up execution leftovers.");
-    boost::lock_guard<boost::mutex> guard(slot_map_mtx_);
-    ROS_DEBUG_STREAM_NAMED(name_, "Exiting run method with goal status "
-                           << (int)concurrency_slots_[slot].goal_handle.getGoalStatus().status
-                           << ": "<< concurrency_slots_[slot].goal_handle.getGoalStatus().text);
-    threads_.remove_thread(concurrency_slots_[slot].thread_ptr);
-    delete concurrency_slots_[slot].thread_ptr;
-    concurrency_slots_.erase(slot);
-    execution_ptr->postRun();
+    slot.execution->join();
+    ROS_DEBUG_STREAM_NAMED(name_, "Execution completed with goal status "
+                           << (int)slot.goal_handle.getGoalStatus().status << ": "<< slot.goal_handle.getGoalStatus().text);
+    slot.execution->postRun();
+    slot.in_use = false;
   }
 
   virtual void reconfigureAll(
@@ -154,7 +170,7 @@ class AbstractAction
 
 protected:
   const std::string &name_;
-  const RobotInformation &robot_info_;
+  const mbf_utility::RobotInformation &robot_info_;
 
   RunMethod run_;
   boost::thread_group threads_;
@@ -166,4 +182,4 @@ protected:
 
 }
 
-#endif //MBF_ABSTRACT_NAV__ABSTRACT_ACTION_H_
+#endif /* MBF_ABSTRACT_NAV__ABSTRACT_ACTION_BASE_H_ */